Created world systems
World system is a way for running functions for a specific world state.
A system has two parameters: a world_in and a world_out. It should
use the data in world_in to produce an output and put it in world_out.
Each system should preferably be a pure function. Using the
box_add_system function you can create a "chain" of system that
will run when the box_tick_world function is called.

This system allow a more well structured and clean program that can be
organized around them.

There might be some performance issues in the way this project is
structured right now due to a lot of unecessary
copying/clearing/writing/allocating buffer, but they're not a problem
right now.