Removed the systems and createad a (basic)falling sand simulation
Turns out that systems thing was actually not that good to work with
when you actually start using it for things. It could totally be used
but was just an unecessary pain and doing the things in a simpler way is
just so much easier and more efficient. Right now each system is just a
function that has the world as a parameter and do it' things on it.

Other than that I also started implementing the falling sand simulation
and it was surprisingly easy. It's really cool to see thing like this
falling sand and the game of life simulation that create amazing visual
result with extremely simple rules. Also props to the noita's dev for
he's talk about the tech used in the game, it was quite usefull to be
honest and you should totally take a look at it:
https://www.gdcvault.com/play/1025695/Exploring-the-Tech-and-Design.